test: collect membership related tests into one integration test crate: tests/membership/
